backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.23
PyMySQL==1.1.0
python-dotenv==1.0.0
Flask-Migrate==4.0.5
PyJWT==2.8.0
pandas==2.0.3
openpyxl==3.1.2